Reorganizing internal configuration
