+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/configuration/instrumentation/supported/ScreenRenderingConfiguration.java
@@ -18,7 +18,7 @@
  */
 package co.elastic.apm.android.sdk.configuration.instrumentation.supported;
 
-import co.elastic.apm.android.sdk.configuration.FeatureConfiguration;
+import co.elastic.apm.android.sdk.internal.configuration.FeatureConfiguration;
 
 public class ScreenRenderingConfiguration extends FeatureConfiguration {
     private final boolean enabled;
