+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/configuration/FeatureConfiguration.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.configuration;
+package co.elastic.apm.android.sdk.internal.configuration;
 
 public abstract class FeatureConfiguration {
